eliminated containers folder, moved everything in nested folders inside of components folder
